+++ tiled/server/router.py
@@ -1945,7 +1945,7 @@ def get_router(
             drop_revision=drop_revision,
         )
 
-        response_data = {"id": entry.key}
+        response_data = {"id": entry.node.key}
         if metadata_modified:
             response_data["metadata"] = metadata
         if access_blob_modified:
@@ -2025,7 +2025,7 @@ def get_router(
             drop_revision=drop_revision,
         )
 
-        response_data = {"id": entry.key}
+        response_data = {"id": entry.node.key}
         if metadata_modified:
             response_data["metadata"] = metadata
         if access_blob_modified: